Prod5 v10

+++ analysis/CTA.TMVA.qsub_train.sh
@@ -6,56 +6,49 @@
 #
 
 RPARA=RUNPARA
-let "EBIN = $SGE_TASK_ID - 1"
+NENE=NBINSNBINS
+#let "EBIN = $SGE_TASK_ID - 1"
 
 ulimit -n 2056
 
 # set the right observatory (environmental variables)
 source "${EVNDISPSYS}"/setObservatory.sh CTA
 
-PFIL=${RPARA}_${EBIN}
-rm -f $PFIL.log
+for ((EBIN=0; EBIN < $NENE; EBIN++))
+do
+    PFIL=${RPARA}_${EBIN}
+    rm -f $PFIL.log
 
-# this file is deleted after successful 
-# completion of training
-touch ${PFIL}.${SGE_JOB_ID}.${SGE_TASK_ID}.RUNNING
+    echo ${PFIL}.runparameter
 
-echo ${PFIL}.runparameter
+    echo "JOB ID ${SGE_JOB_ID}" > "${PFIL}".log
+    ${EVNDISPSYS}/bin/trainTMVAforGammaHadronSeparation "${PFIL}".runparameter >> "${PFIL}".log
 
-${EVNDISPSYS}/bin/trainTMVAforGammaHadronSeparation "${PFIL}".runparameter > "${PFIL}".log
+    CDIR=$(dirname "$PFIL".log)
+    # remove .C files (never used; we use the XML files)
+    rm -f "$CDIR"/BDT_"${EBIN}"*.C
+    # remove complete_BDTroot at the end of the run
+    # (generally not used, but takes up lots of disk space)
+    # rm -rf $CDIR/complete_BDTroot/BDT_${EBIN}*
+    rm -rf "$CDIR"/complete_BDTroot
 
-CDIR=$(dirname "$PFIL".log)
-# remove .C files (never used; we use the XML files)
-rm -f "$CDIR"/BDT_"${EBIN}"*.C
-# remove complete_BDTroot at the end of the run
-# (generally not used, but takes up lots of disk space)
-# rm -rf $CDIR/complete_BDTroot/BDT_${EBIN}*
-rm -rf "$CDIR"/complete_BDTroot
-
-# check successful completion of training
-# remove temporary file
-if [ -e ${PFIL}.log ]; then
-   TSTRING=$(tail -n 1 ${PFIL}.log | grep Complete)
-   echo "$TSTRING"
-   if [ ! -z "$TSTRING" ]; then
-      rm -f ${PFIL}.${SGE_JOB_ID}.${SGE_TASK_ID}.RUNNING
-   fi
-   TSTRING=$(tail -n 3 ${PFIL}.log | grep "not enough")
-   if [ ! -z "$TSTRING" ]; then
-      rm -f ${PFIL}.${SGE_JOB_ID}.${SGE_TASK_ID}.RUNNING
-   fi
-fi
-
-# mv log file into root file
-if [ -e ${PFIL}.log ]
-then
-    ${EVNDISPSYS}/bin/logFile tmvaLog $CDIR/BDT_${EBIN}.root ${PFIL}.log
-    rm -f ${PFIL}.log
-fi
-if [ -e ${PFIL}.runparameter ] && [ -e $CDIR/BDT_${EBIN}.root ]
-then
-    ${EVNDISPSYS}/bin/logFile tmvaRunparameter $CDIR/BDT_${EBIN}.root ${PFIL}.runparameter
-    rm -f ${PFIL}.runparameter
-fi
+    # mv log file into root file
+    if [ -e ${PFIL}.log ] && [ -e $CDIR/BDT_${EBIN}.root ]
+    then
+        ${EVNDISPSYS}/bin/logFile tmvaLog $CDIR/BDT_${EBIN}.root ${PFIL}.log
+        rm -f ${PFIL}.log
+    fi
+    if [ -e ${PFIL}.runparameter ] && [ -e $CDIR/BDT_${EBIN}.root ]
+    then
+        ${EVNDISPSYS}/bin/logFile tmvaRunparameter $CDIR/BDT_${EBIN}.root ${PFIL}.runparameter
+        rm -f ${PFIL}.runparameter
+    fi
+    # mv xml file into root file
+    if [ -e $CDIR/BDT_${EBIN}_BDT_0.weights.xml ] && [ -e $CDIR/BDT_${EBIN}.root ]
+    then
+       ${EVNDISPSYS}/bin/logFile tmvaXML $CDIR/BDT_${EBIN}.root $CDIR/BDT_${EBIN}_BDT_0.weights.xml
+       rm -f $CDIR/BDT_${EBIN}_BDT_0.weights.xml
+    fi
+done
 
 exit
